% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTheoModelFun.R
\name{eval_coex}
\alias{eval_coex}
\title{eval_coex}
\usage{
eval_coex(list_int_coex, Nrun = 1000, vec_p_d = seq(0.001, 0.5, length =
  20), vec_p_s = seq(0.001, 0.5, length = 20), vec_K = c(1, 5, 10, 100))
}
\arguments{
\item{Nrun}{Number of run of cellular automaton}

\item{p_d}{vector of mortality rate to explore}

\item{p_s}{vector of succession rate to explore}

\item{K}{strength of the competitive asymmetry parameter}

\item{list_int}{list of initial landscape}
}
\value{
list of landscape results
}
\description{
.. Run simulation of cellur automaton for all combination mortality rate and succession rate to evalute their effect on species coexistence ..
}
\details{
.. content for \details{} ..
}
\author{
Georges Kunstler
}

