% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTheoModelFun.R
\name{eval_converg}
\alias{eval_converg}
\title{eval_converg}
\usage{
eval_converg(list_init, Niter, Nrun, p_d, p_s, K)
}
\arguments{
\item{list_init}{Initila list of the landscape}

\item{Niter}{number of iteration}

\item{Nrun}{number of cellular automaton update run per iteration}

\item{p_d}{mortality rate paramter}

\item{p_s}{succession rate paramter}

\item{K}{strength of the competitive asymmetry parameter}
}
\value{
list of landscape list
}
\description{
.. Run sequence of Nrun simulation Niter times to explore convergence. ..
}
\details{
.. content for \details{} ..
}
\author{
Georges Kunstler
}

