% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTheoModelFun.R
\name{GenerateRandSp}
\alias{GenerateRandSp}
\title{GenerateRandSp}
\usage{
GenerateRandSp(nsp, Nval = 1000)
}
\arguments{
\item{nsp}{number of species}

\item{Nval}{number of value to drwan from (must be greater than nsp)}
}
\value{
data.frame of the three species parameters (E, L, S)
}
\description{
.. Create the data.frame describing a set of nsp species with a triangular constrain on their three parameters (early successional competitive ability, late successional competitive ability, abiotic stress tolerance). ..
}
\details{
.. Generate nsp equally spaced species in traiangular space of three parameters (early successional competitive ability, late successional competitive ability, abiotic stress tolerance) representing teh species strategies. These species are randomly drawn from Nval value of the parameters. ..
}
\author{
Georges Kunstler
}

