% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTheoModelFun.R
\name{image_landscape}
\alias{image_landscape}
\title{image_landscape}
\usage{
image_landscape(list_res, c_e, c_l, c_s, ...)
}
\arguments{
\item{list_res}{landcsape list}

\item{c_e}{vector of species parameters c}

\item{c_l}{vector of species parameters l}

\item{c_s}{vector of species parameters s}
}
\value{
plot
}
\description{
.. Plot RGB of species three parameters over the landscape (one color for each parameter). ..
}
\details{
.. content for \details{} ..
}
\author{
Georges Kunstler
}

