% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTheoModelFun.R
\name{image_landscape_e}
\alias{image_landscape_e}
\title{image_landscape_e}
\usage{
image_landscape_e(list_res, c_)
}
\arguments{
\item{list_res}{landscape list}

\item{c_}{species parameters}
}
\value{
plot
}
\description{
.. Plot a landscape list with one species parameters. ..
}
\details{
.. content for \details{} ..
}
\author{
Georges Kunstler
}

