% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTheoModelFun.R
\name{image_return_mat}
\alias{image_return_mat}
\title{image_return_mat}
\usage{
image_return_mat(m, ...)
}
\arguments{
\item{m}{matrix}
}
\value{
plot
}
\description{
.. plot matrix as a raster..
}
\details{
.. content for \details{} ..
}
\author{
Georges Kunstler
}

