% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTheoModelFun.R
\name{InitLandscape}
\alias{InitLandscape}
\title{InitLandscape}
\usage{
InitLandscape(df_sp, NN = 100, Nlandscape = 1, min_ss = 0, max_ss = 0,
  per_occup = 0.5)
}
\arguments{
\item{df_sp}{data.frame with species parameter}

\item{NN}{size of square landscape}

\item{Nlandscape}{number of time the landscape is repeated}

\item{min_ss}{minimun abiotic stress value}

\item{max_ss}{maximum abiotic stress value}

\item{per_occup}{percentage of cell occuped by a species}
}
\value{
list with first element a matrix (NN, NN*Nlandscape) with species number (-100 no species), second element matrix off successional stage of the cells, third element a vector of teh abiotic gradient stress (size NN*Nlandscape), fourst element species early successional competitive ability, fiveth element species late successional competitive ability, sixt element species abiotic stress tolerance.
}
\description{
.. Generate initial landscape for cellular automaton simulation ..
}
\details{
.. Take a data.frame of species parameters and generate initial condition with random assignment of specie sto cells with a given size of the landscape (defined as square landscape of size NN repeated Nlandscape times to cover the abiotic stress gradient). The abiotic stress gradient is linear form min_ss to max_ss and run from along NN*Nlandscape. ..
}
\author{
Georges Kunstler
}

