% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTheoModelFun.R
\name{return_mat_fill_c_raster}
\alias{return_mat_fill_c_raster}
\title{return_mat_fill_c_raster}
\usage{
return_mat_fill_c_raster(m, cc)
}
\arguments{
\item{m}{matrix with species number}

\item{cc}{species parameters value}
}
\value{
raster
}
\description{
.. Fill the matrix with species parameters adn return it as a raster object ..
}
\details{
.. content for \details{} ..
}
\author{
Georges Kunstler
}

