% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTheoModelFun.R
\name{return_mat_raster}
\alias{return_mat_raster}
\title{return_mat_raster}
\usage{
return_mat_raster(m)
}
\arguments{
\item{m}{matrix}
}
\value{
raster
}
\description{
.. Convert matrix to raster for plot ..
}
\details{
.. content for \details{} ..
}
\author{
Georges Kunstler
}

