% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTheoModelFun.R
\name{table_level}
\alias{table_level}
\title{table_level}
\usage{
table_level(x, nsp)
}
\arguments{
\item{x}{vector or matrix of species number}

\item{nsp}{number of species}
}
\description{
.. Compute abundance of each species. ..
}
\details{
.. content for \details{} ..
}
\author{
Georges Kunstler
}

