% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTheoModelFun.R
\name{TernPlotSp}
\alias{TernPlotSp}
\title{TernPlotSp}
\usage{
TernPlotSp(Nval = 1000, add_text = FALSE)
}
\arguments{
\item{Nval}{number of species used to generate the plot.}
}
\value{
plot
}
\description{
.. Create a plot of species three parameters early successional competitive ability, late successional competitive ability, abiotic stress tolerance..
}
\details{
.. content for \details{} ..
}
\author{
Georges KUnstler
}

