% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateIterR}
\alias{UpdateIterR}
\title{Run cellular automaton based on initila condition and species parameters n times}
\usage{
UpdateIterR(mat_sp, mat_suc, c_e, c_l, c_s, ss, prob_distur, prob_suc, K, n)
}
\arguments{
\item{mat_sp}{landscape matrix of cells with species code if occupied.}

\item{c_e}{vector of species early successional competitive ability.}

\item{c_l}{vector of species late successional competitive ability.}

\item{c_s}{vector of species abiotic stress tolerance.}

\item{ss}{vector of abiotic stress.}

\item{prob_distur}{probablity of cell disturbance (mortality).}

\item{prob_suc}{probablity of cell succession.}

\item{K}{strength of the competitive asymmetry parameter.}

\item{n}{number of repetition of landscape updating.}

\item{mat_succ}{landscape matrix of cells successional status.}
}
\description{
Run cellular automaton based on initila condition and species parameters n times
}

